pyproject.toml
[tool.pdm]

[project]
name = "drf-template"
version = "0.1.0"
description = ""
authors = [
    {name = "devsarvesh92", email = "sarveshdev92@gmail.com"},
]
dependencies = [
    "django>=4.1.6",
    "djangorestframework>=3.14.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"
